backend/requirements.txt
asgiref==3.3.4
DateTime==4.3
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
future==0.18.2
pytz==2021.1
serializers==0.2.4
sqlparse==0.4.1
zope.interface==5.4.0
